+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ##### Enhancements
 
-* Add support for pre/post-actions in scheme actions
+* Add support for pre/post-actions in scheme actions  
   [thiagohmcruz](https://github.com/thiagohmcruz)
   [#401](https://github.com/CocoaPods/CocoaPods/issues/401)